refactor(runtime): extract turn/redact.rs into agent/turn — #7415
