+++ crates/zeroclaw-runtime/src/agent/turn/mod.rs
@@ -1,0 +1,10 @@
+//! The agent turn engine, decomposed into single-purpose step modules.
+//!
+//! Public paths are unchanged: external code keeps importing via
+//! `crate::agent::loop_::*` (re-export block there). See the run sheet in
+//! `/opt/notes/work/zeroclaw/unification_modular/RUN_SHEET.md` during the
+//! #7415 migration.
+
+pub(crate) mod redact;
+
+pub use redact::scrub_credentials;
